Unzip last modified zip

A function to download and unzip the last modified zip file from a specified Google Drive folder.

Usage

unzip_last_modified_zip(folder_id, exdir = tempdir())

Arguments

folder_id

The ID of the googledrive folder to check for the last modified zip file.

exdir

The directory where the zip file should be extracted. Defaults to tempdir().

Value

The path to the datapackage.json of the last modified zip file in the specified folder.

See also

Examples

if (FALSE) { # \dontrun{
datapackage <- unzip_last_modified_zip(folder_id = "17p2MZt9LIuhIU72u_JjDBbO7D1IPTv7-") |>
camtraptor::read_camtrap_dp()
} # }